[REQUESTS] Fix bugs

+++ mpi-backend/src/main/java/itmo/mpi/repository/IslandRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IslandRepository extends JpaRepository<Island, Long> {
+public interface IslandRepository extends JpaRepository<Island, Integer> {
 }
